FIX: adds the PubSub's Message ID to the Headers
The Message ID is the ID of this message,
assigned by the server when the message is published.
Guaranteed to be unique within the topic.
This value may be read by a subscriber that receives a
PubsubMessage via a subscriptions.pull call or a push delivery.
It must not be populated by the publisher in a topics.publish call.

more info:
https://cloud.google.com/pubsub/docs/reference/rest/v1/PubsubMessage